test projects list

+++ api/index.php
@@ -331,8 +331,8 @@ $route->get(array("/api/user/me/projects",
                   "/api/project/list"), function($id = null){
 
     $id = ($id === null) ? force_auth() : (int) $id;
-    $offset = get("offset",true) || 0;
-    $number = get("number",true) || 20;
+    $offset = ((int) get("offset",true)) || 0;
+    $number = ((int) get("number",true)) || 20;
 
     $list = get_projects_for_user($id, $offset, $number);
     $output = array("total" => count($list),
